It finally works!!!
Yeet!

I had to make the pieces lock into the endPosition once they got within 0.1f of it, because the Pawn was stopping like 0.01f away and that was messing up the comparison. But I'm pretty sure it works now so yay! Also added a method to show all pieces under threat from the opponent in GameManager.cs, to enable this, tick the ShouldShowThreats setting in the GameManager inspector.